🐛 [#109] - Fix PR review issues for dev debug login 🔧
- 🔧 Fixed `(bool) env()` cast in devlogin.php — now uses `=== true` strict comparison so only literal `true` activates the feature
- 🔒 Normalized allowed environments to lowercase in DevLoginGuard to block case variants like `Production`
- 📦 Fixed roles response in ListDevUsersController — now returns `string[]` via `pluck('name')` matching frontend DevUser type
- 🔇 Fixed silent error swallowing in dev-api.ts — only silences 404 (feature disabled), rethrows other errors
- 🪝 Extracted all state/effects/handlers from DevDebugger into `use-dev-debugger.ts` hook (CLAUDE.md custom hook convention)
- 🙈 Gated Dev Login section on `devLoginSectionVisible` so it hides when backend returns null (feature disabled)
- 🧪 Fixed DevDebugger.test.tsx mock for `@tanstack/react-query` to include `useQuery`
- ⚡ Removed `vi.resetModules()` test antipattern by moving `SHOULD_START_HIDDEN` into useState initializer
