+++ code/webapp/cypress/e2e/dev-debug-login.cy.ts
@@ -18,11 +18,11 @@
  */
 
 before(() => {
-  cy.task('test:reset', null, { timeout: 60_000 })
+    cy.task('test:reset', null, { timeout: 60_000 })
 })
 
 beforeEach(() => {
-  cy.clearLocalStorage()
+    cy.clearLocalStorage()
 })
 
 // ══════════════════════════════════════════════════════════════════════════════
@@ -30,53 +30,53 @@ beforeEach(() => {
 // ══════════════════════════════════════════════════════════════════════════════
 
 describe('Dev Debug Login — Happy Path', () => {
-  it('muestra la sección Dev Login en el DevDebugger cuando no hay sesión', () => {
-    cy.visit('/login')
+    it('muestra la sección Dev Login en el DevDebugger cuando no hay sesión', () => {
+        cy.visit('/login')
 
-    // Show DevDebugger (hidden by default in e2e — trigger with Ctrl+Shift+D)
-    cy.get('body').type('{ctrl}{shift}d')
+        // Show DevDebugger (hidden by default in e2e — trigger with Ctrl+Shift+D)
+        cy.get('body').type('{ctrl}{shift}d')
 
-    cy.contains('Dev Debugger', { timeout: 5_000 }).should('be.visible')
-    cy.contains('Dev Login').should('be.visible')
-  })
-
-  it('lista usuarios y permite iniciar sesión sin contraseña', () => {
-    cy.visit('/login')
-
-    cy.get('body').type('{ctrl}{shift}d')
-    cy.contains('Dev Login', { timeout: 5_000 }).should('be.visible')
-
-    // Users should appear (seeded by test:reset)
-    cy.get('[data-testid="dev-login-user"]', { timeout: 10_000 })
-      .should('have.length.at.least', 1)
-      .first()
-      .click()
-
-    // After clicking, should be authenticated and redirected away from login
-    cy.url({ timeout: 15_000 }).should('not.include', '/login')
-
-    // Auth state persisted in localStorage
-    cy.window().then((win) => {
-      const raw = win.localStorage.getItem('auth-storage')
-      expect(raw).to.exist
-      const { state } = JSON.parse(raw!)
-      expect(state.token).to.exist
-      expect(state.isAuthenticated).to.be.true
+        cy.contains('Dev Debugger', { timeout: 5_000 }).should('be.visible')
+        cy.contains('Dev Login').should('be.visible')
     })
-  })
 
-  it('filtra usuarios por nombre con el buscador local', () => {
-    cy.visit('/login')
+    it('lista usuarios y permite iniciar sesión sin contraseña', () => {
+        cy.visit('/login')
 
-    cy.get('body').type('{ctrl}{shift}d')
-    cy.contains('Dev Login', { timeout: 5_000 }).should('be.visible')
+        cy.get('body').type('{ctrl}{shift}d')
+        cy.contains('Dev Login', { timeout: 5_000 }).should('be.visible')
 
-    // Wait for list to load
-    cy.get('[data-testid="dev-login-user"]', { timeout: 10_000 }).should('exist')
+        // Users should appear (seeded by test:reset)
+        cy.get('[data-testid="dev-login-user"]', { timeout: 10_000 })
+            .should('have.length.at.least', 1)
+            .first()
+            .click()
 
-    // Type something that should filter to 0 or fewer results
-    cy.get('[placeholder="Buscar usuario..."]').type('zzznoresult')
-    cy.contains('Sin resultados').should('be.visible')
-    cy.get('[data-testid="dev-login-user"]').should('not.exist')
-  })
+        // After clicking, should be authenticated and redirected away from login
+        cy.url({ timeout: 15_000 }).should('not.include', '/login')
+
+        // Auth state persisted in localStorage
+        cy.window().then((win) => {
+            const raw = win.localStorage.getItem('auth-storage')
+            expect(raw).to.exist
+            const { state } = JSON.parse(raw!)
+            expect(state.token).to.exist
+            expect(state.isAuthenticated).to.be.true
+        })
+    })
+
+    it('filtra usuarios por nombre con el buscador local', () => {
+        cy.visit('/login')
+
+        cy.get('body').type('{ctrl}{shift}d')
+        cy.contains('Dev Login', { timeout: 5_000 }).should('be.visible')
+
+        // Wait for list to load
+        cy.get('[data-testid="dev-login-user"]', { timeout: 10_000 }).should('exist')
+
+        // Type something that should filter to 0 or fewer results
+        cy.get('[placeholder="Buscar usuario..."]').type('zzznoresult')
+        cy.contains('Sin resultados').should('be.visible')
+        cy.get('[data-testid="dev-login-user"]').should('not.exist')
+    })
 })
